Move mirage models into mirage/models
